- Summary: 229538: Big number display issue in chart builder. 310253: 	Missing quarter unit in scale interval.
 - Bugzilla Bug (s) Resolved: 229538, 310253

 - Description:
229538: Fixed. Now the bar, line, area, pie, meter, scatter, bubble, difference, tube, cone and  pyramid charts are support big number, stock and gantt no need to support big number.
310253: Fixed, added Quarter in scale units.

- Tests Description:  Manual Test

- Notes to Build Team: N/A

- Notes to Developers: N/A

- Notes to QA:  N/A

- Notes to Documentation:  N/A

- Files Edited:
@changelist

- Files Added:
@addlist

- Files Deleted
@deletelist

+++ chart/org.eclipse.birt.chart.engine/src/org/eclipse/birt/chart/computation/DataPointHints.java
@@ -554,7 +554,7 @@ public final class DataPointHints
 		DecimalFormat df = null;
 		if ( fs == null && oBaseValue instanceof Number )
 		{
-			df = new DecimalFormat( ValueFormatter.getNumericPattern( ( (Number) oBaseValue ).doubleValue( ) ) );
+			df = new DecimalFormat( ValueFormatter.getNumericPattern( (Number) oBaseValue ) );	
 		}
 		
 		try
